+++ lib/core/enums/turbo_tag_type.dart
@@ -1,0 +1,10 @@
+enum TurboTagType {
+  workspace;
+
+  String get argument {
+    switch (this) {
+      case TurboTagType.workspace:
+        return name;
+    }
+  }
+}